+++ src/Core/Credentials/AuthToken.php
@@ -16,14 +16,36 @@ namespace Bitrix24\SDK\Core\Credentials;
 use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
 use Symfony\Component\HttpFoundation;
 
-class AuthToken
+readonly final class AuthToken
 {
+    /**
+     * @throws InvalidArgumentException
+     */
     public function __construct(
-        protected string  $accessToken,
-        protected ?string $refreshToken,
-        protected int     $expires,
-        protected ?int    $expiresIn = null)
+        /**
+         * @phpstan-param non-empty-string $accessToken
+         */
+        public string  $accessToken,
+        /**
+         * @phpstan-param non-empty-string|null $refreshToken
+         */
+        public ?string $refreshToken,
+        /**
+         * @phpstan-param non-negative-int $expires
+         */
+        public int     $expires,
+        /**
+         * @phpstan-param non-negative-int|null $expiresIn
+         */
+        public ?int    $expiresIn = null)
     {
+        if (trim($this->accessToken) === '') {
+            throw new InvalidArgumentException('accessToken cannot be empty string');
+        }
+
+        if ($this->refreshToken !== null && trim($this->refreshToken) === '') {
+            throw new InvalidArgumentException('refreshToken cannot be empty string');
+        }
     }
 
     /**
@@ -36,42 +58,67 @@ class AuthToken
         return $this->refreshToken === null;
     }
 
-    public function getAccessToken(): string
-    {
-        return $this->accessToken;
-    }
-
-    public function getRefreshToken(): ?string
-    {
-        return $this->refreshToken;
-    }
-
-    public function getExpires(): int
-    {
-        return $this->expires;
-    }
-
     public function hasExpired(): bool
     {
-        return $this->getExpires() <= time();
+        return $this->expires <= time();
     }
 
-
-    public static function initFromArray(array $request): self
+    /**
+     * Initialize an AuthToken object from an array of token payload.
+     *
+     * @param array $authTokenPayload The array containing the token payload.
+     *   - access_token (string): The access token.
+     *   - refresh_token (string): The refresh token.
+     *   - expires (int): The expiration timestamp of the token.
+     *
+     * @return self The initialized AuthToken object.
+     *
+     * @throws InvalidArgumentException If any of the required fields are not found in the authTokenPayload array.
+     */
+    public static function initFromArray(array $authTokenPayload): self
     {
+        if (!array_key_exists('access_token', $authTokenPayload)) {
+            throw new InvalidArgumentException('field access_token not fount in authTokenPayload');
+        }
+
+        if (!array_key_exists('refresh_token', $authTokenPayload)) {
+            throw new InvalidArgumentException('field refresh_token not fount in authTokenPayload');
+        }
+
+        if (!array_key_exists('expires', $authTokenPayload)) {
+            throw new InvalidArgumentException('field expires not fount in authTokenPayload');
+        }
+
         return new self(
-            (string)$request['access_token'],
-            (string)$request['refresh_token'],
-            (int)$request['expires']
+            (string)$authTokenPayload['access_token'],
+            (string)$authTokenPayload['refresh_token'],
+            (int)$authTokenPayload['expires']
         );
     }
 
+    /**
+     * Initializes an object of this class from a Workflow Request object.
+     *
+     * @param \Symfony\Component\HttpFoundation\Request $request The Workflow Request object.
+     *
+     * @return self The initialized object of this class.
+     */
     public static function initFromWorkflowRequest(HttpFoundation\Request $request): self
     {
         $requestFields = $request->request->all();
         return self::initFromArray($requestFields['auth']);
     }
 
+    /**
+     * Initializes the object from an event request.
+     *
+     * Parses the provided HTTP request and extracts the required fields to construct a new instance of the current class.
+     * The required fields are 'access_token', 'expires', and 'expires_in'.
+     *
+     * @param \Symfony\Component\HttpFoundation\Request $request The HTTP request object containing the required fields.
+     *
+     * @return self Returns a new instance of the current class with the extracted data.
+     */
     public static function initFromEventRequest(HttpFoundation\Request $request): self
     {
         $requestFields = $request->request->all();
@@ -84,7 +131,11 @@ class AuthToken
     }
 
     /**
-     * @throws InvalidArgumentException
+     * Initializes an instance of the class from a placement request.
+     *
+     * @param HttpFoundation\Request $request The placement request object.
+     * @return self The initialized instance of the class.
+     * @throws InvalidArgumentException If the required fields are not found in the request.
      */
     public static function initFromPlacementRequest(HttpFoundation\Request $request): self
     {